Item system overhaul. Bug fixing. New whens.
Completely changed the item system, for allowing different "ItemInfo" per when. Changed the item nbt datas. Fixed a bug where player who receive the item, will also have the effect applied. Removed "arrow" when and changed it with "projectile: when, with support for enderpearls and snowballs.  Added armorequip and armorunequip when.